# openapi.yaml
openapi: 3.0.0
info:
  version: 0.2.3
  title: Owncast APIs
  description: |-
    Internal
    The APIs required to make the Owncast web interface and internal functionality work. These APIs will never reach stability and will always change as Owncast features and functionality change. This means these endpoints are not good options to build your own tooling and features on top of. You can, however, of course can if that risk is acceptable to you.

    Admin
    The APIs used to configure the Owncast server, as well as power the web admin interface. These APIs are protected by the admin password, using HTTP Basic authentication. They should not be used by external tools or integrations as it would require you to give your admin password to a third party.
  contact:
    name: Gabe Kangas
    email: gabek@real-ity.com
    url: "http://owncast.online"
  license:
    url: "https://github.com/Space-Software-LTDA/owncast/blob/develop/LICENSE"
    name: MIT
  x-logo:
    url: >-
...
servers:
  - url: "http://localhost:8080/api"
    description: Localhost port 8080
paths:
...
  /chat:
    get:
      summary: Gets a list of chat messages
      operationId: GetChatMessages
      tags: ["Internal", "Chat"]
      parameters:
        - $ref: "#/components/parameters/AccessToken"
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessages"
        "401":
          $ref: "#/components/responses/401"
        "501":
          $ref: "#/components/responses/501"
  /chat/register:
    post:
      summary: Registers an anonymous chat user
      operationId: RegisterAnonymousChatUser
      tags: ["Internal", "Chat"]
      parameters:
        - in: header
          name: X-Forwarded-User
          schema:
            type: string
          required: false
      requestBody:
        description: The display name of the new user
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                email:
                  type: string
      responses:
        "200":
          description: Anonymous user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnonymousUser"
        "400":
          $ref: "#/components/responses/400"
    options:
      operationId: RegisterAnonymousChatUserOptions
      x-internal: true
      tags: ["Objects", "Chat"]
      responses:
        "204":
          $ref: "#/components/responses/204Options"
...
  /admin/chat/clients:
    get:
      summary: Get a detailed list of currently connected chat clients
      operationId: GetConnectedChatClients
      tags: ["Internal", "Admin", "Chat"]
      security:
        - BasicAuth: []
      responses:
        "200":
          description: List of chat clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatClients"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401BasicAuth"
        default:
          $ref: "#/components/responses/Default"
    options:
      operationId: GetConnectedChatClientsOptions
      x-internal: true
      tags: ["Objects", "Chat"]
      responses:
        "204":
          $ref: "#/components/responses/204"
...
  /integrations/chat/system/client/{clientId}:
    post:
      summary: Send a system message to a single client
      operationId: SendSystemMessageToConnectedClient
      tags: ["External", "Chat"]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: integer
          description: The ID of the specified client
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemMessage"
      responses:
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
        default:
          $ref: "#/components/responses/Default"
    options:
      operationId: SendSystemMessageToConnectedClientOptions
      x-internal: true
      tags: ["Objects", "Internal", "Admin", "Chat"]
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: integer
          description: The ID of the specified client
      responses:
        "204":
          $ref: "#/components/responses/204"
...
  /integrations/chat:
    get:
      summary: Get chat history
      operationId: ExternalGetChatMessages
      tags: ["External", "Chat"]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessages"
        "401":
          $ref: "#/components/responses/401"
        "501":
          $ref: "#/components/responses/501"
        default:
          $ref: "#/components/responses/Default"
    options:
      operationId: ExternalGetChatMessagesOptions
      x-internal: true
      tags: ["Objects", "External", "Admin", "Chat"]
      responses:
        "204":
          $ref: "#/components/responses/204"
  /integrations/clients:
    get:
      summary: Connected clients
      operationId: ExternalGetConnectedChatClients
      tags: ["External", "Chat"]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of chat clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatClients"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401BasicAuth"
        default:
          $ref: "#/components/responses/Default"
    options:
      operationId: ExternalGetConnectedChatClientsOptions
      x-internal: true
      tags: ["Objects", "External", "Admin", "Chat"]
      responses:
        "204":
          $ref: "#/components/responses/204"
...
  /auth/indieauth:
    post:
      summary: Begins auth flow
      operationId: StartIndieAuthFlow
      tags: ["Internal", "Auth", "Chat"]
      parameters:
        - $ref: "#/components/parameters/AccessToken"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                authHost:
                  type: string
      responses:
        "200":
          description: Authentication redirect
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect:
                    type: string
                    format: url
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        default:
          $ref: "#/components/responses/Default"
...
  /auth/provider/indieauth:
    get:
      summary: Handles the IndieAuth auth endpoint
      operationId: HandleIndieAuthEndpointGet
      tags: ["Internal", "Auth", "Chat"]
      security:
        - BasicAuth: []
      parameters:
        - $ref: "#/components/parameters/IndieAuthClientId"
        - $ref: "#/components/parameters/IndieAuthRedirectURI"
        - $ref: "#/components/parameters/IndieAuthCodeChallenge"
        - $ref: "#/components/parameters/IndieAuthState"
      responses:
        "200":
          description: IndieAuth flow concluded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndieAuthResponse"
        "307":
          description: Redirected to redirect URI
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
    post:
      summary: Handles IndieAuth from form submission
      operationId: HandleIndieAuthEndpointPost
      tags: ["Internal", "Auth", "Chat"]
      requestBody:
        $ref: "#/components/requestBodies/IndieAuthEndpointForm"
      responses:
        "200":
          description: IndieAuth flow concluded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/IndieAuthResponse"
                  - $ref: "#/components/schemas/IndieAuthServerProfileResponse"
        "400":
          $ref: "#/components/responses/400"
  /auth/fediverse:
    post:
      summary: Register a Fediverse OTP request
      operationId: RegisterFediverseOTPRequest
      tags: ["Internal", "Auth", "Chat"]
      parameters:
        - $ref: "#/components/parameters/AccessToken"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                account:
                  type: string
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseAPIResponse"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
  /auth/fediverse/verify:
    post:
      summary: Verify Fediverse OTP code
      operationId: VerifyFediverseOTPRequest
      tags: ["Internal", "Auth", "Chat"]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        "200":
          description: OTP Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseAPIResponse"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
components:
  schemas:
    Error:
      type: object
      description: Structure for an error response
      properties:
        error:
          type: string
    BaseAPIResponse:
      type: object
      description: Simple API response
      properties:
        message:
          type: string
        success:
          type: boolean
...
    AnonymousUser:
      type: object
      properties:
        id:
          type: string
        accessToken:
          type: string
        displayName:
          type: string
    User:
      type: object
      properties:
        createdAt:
          type: string
        disabledAt:
          type: string
        nameChangedAt:
          type: string
        id:
          type: string
        displayName:
          type: string
        previousNames:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        displayColor:
          type: integer
        isBot:
          type: boolean
        authenticated:
          type: boolean
...
    Event:
      type: object
      properties:
        timestamp:
          type: string
        type:
          type: string
        id:
          type: string
    UserEvent:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        hiddenAt:
          type: string
        clientId:
          type: integer
    MessageEvent:
      type: object
      properties:
        body:
          type: string
    ChatMessages:
      type: array
      items:
        anyOf:
          - $ref: "#/components/schemas/UserMessage"
          - $ref: "#/components/schemas/SystemMessage"
          - $ref: "#/components/schemas/ActionMessage"
          - $ref: "#/components/schemas/FederatedAction"
    UserMessage:
      type: object
      allOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/UserEvent"
        - $ref: "#/components/schemas/MessageEvent"
    SystemMessage:
      type: object
      allOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/MessageEvent"
...
    ActionMessage:
      type: object
      allOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/MessageEvent"
    FederatedAction:
      type: object
      allOf:
        - $ref: "#/components/schemas/Event"
        - $ref: "#/components/schemas/MessageEvent"
        - type: object
          properties:
            image:
              type: string
            link:
              type: string
            title:
              type: string
...
    GeoDetails:
      type: object
      properties:
        countryCode:
          type: string
        regionName:
          type: string
        timeZone:
          type: string
...
    ChatClient:
      type: object
      properties:
        connectedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        geo:
          $ref: "#/components/schemas/GeoDetails"
        userAgent:
          type: string
        messageCount:
          type: integer
    ChatClients:
      type: array
      items:
        $ref: "#/components/schemas/ChatClient"
...
    IndieAuthResponse:
      type: object
      properties:
        me:
          type: string
        profile:
          $ref: "#/components/schemas/IndieAuthProfile"
        error:
          type: string
        errorDescription:
          type: string
    IndieAuthProfile:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        photo:
          type: string
    IndieAuthServerProfileResponse:
      type: object
      properties:
        me:
          type: string
        profile:
          $ref: "#/components/schemas/IndieAuthServerProfile"
        error:
          type: string
        errorDescription:
          type: string
    IndieAuthServerProfile:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        photo:
          type: string
  headers:
    "Access-Control-Allow-Origin":
      schema:
        type: string
    "Access-Control-Allow-Credentials":
      schema:
        type: string
        format: boolean
    "Access-Control-Allow-Headers":
      schema:
        type: string
    "WWW-Authenticate":
      schema:
        type: string
    "X-Content-Type-Options":
      schema:
        type: string
  requestBodies:
...
    IndieAuthEndpointForm:
      description: Form submission from IndieAuth
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            properties:
              code:
                type: string
              redirect_uri:
                type: string
              client_id:
                type: string
              codeVerifier:
                type: string
  responses:
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseAPIResponse"
    "204":
      description: No Content
    "204Options":
      description: No Content
      x-internal: true
      headers:
        "Access-Control-Allow-Origin":
          $ref: "#/components/headers/Access-Control-Allow-Origin"
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseAPIResponse"
    "401":
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
    "401BasicAuth":
      description: Basic auth failure
      headers:
        "Access-Control-Allow-Origin":
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        "Access-Control-Allow-Credentials":
          $ref: "#/components/headers/Access-Control-Allow-Credentials"
        "Access-Control-Allow-Headers":
          $ref: "#/components/headers/Access-Control-Allow-Headers"
        "WWW-Authenticate":
          $ref: "#/components/headers/WWW-Authenticate"
        "X-Content-Type-Options":
          $ref: "#/components/headers/X-Content-Type-Options"
      content:
        text/plain:
          schema:
            type: string
    "403":
      description: Forbidden
...
    "500":
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "501":
      description: Not Implemented
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
...
    AccessToken:
      in: query
      name: accessToken
      schema:
        type: string
      required: true
    IndieAuthClientId:
      in: query
      name: client_id
      schema:
        type: string
      required: true
    IndieAuthRedirectURI:
      in: query
      name: redirect_uri
      schema:
        type: string
      required: true
    IndieAuthCodeChallenge:
      in: query
      name: code_challenge
      schema:
        type: string
      required: true
    IndieAuthState:
      in: query
      name: state
      schema:
        type: string
      required: true
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
...